add target resource. fix errors checking for new targets
